split classes to be sipler
